Move: VO 객체들을 관리하기 위해 vo 패키지 추가
