+++ src/main/java/com/sixman/roomus/product/command/domain/model/Product.java
@@ -3,6 +3,7 @@ package com.sixman.roomus.product.command.domain.model;
 import com.sixman.roomus.commons.jpa.MoneyConverter;
 import com.sixman.roomus.commons.model.Money;
 import com.sixman.roomus.product.command.domain.exception.NotProductOwnerException;
+import com.sixman.roomus.product.command.domain.model.vo.ProductScale;
 import lombok.*;
 
 import javax.persistence.*;
