+++ src/main/java/com/sixman/roomus/rooms/command/application/controller/RoomController.java
@@ -20,7 +20,7 @@ import java.util.List;
 @RestController
 @RequestMapping("/v1/rooms")
 @RequiredArgsConstructor
-@Tag(name = "방", description = "방에 관련된 조회를 제외한 API")
+@Tag(name = "방", description = "방 관련 CUD API")
 public class RoomController {
 
     private final RoomService roomService;
