+++ src/main/java/com/sixman/roomus/rooms/command/domain/model/vo/Float4.java
@@ -1,11 +1,8 @@
-package com.sixman.roomus.rooms.command.domain.model;
+package com.sixman.roomus.rooms.command.domain.model.vo;
 
 import lombok.*;
 import org.hibernate.annotations.ColumnDefault;
-import org.hibernate.annotations.DynamicInsert;
-import org.hibernate.annotations.DynamicUpdate;
 
-import javax.persistence.Column;
 import javax.persistence.Embeddable;
 import java.io.Serializable;
 import java.util.Objects;
