+++ src/main/java/com/sixman/roomus/rooms/command/domain/model/vo/RoomLikesMemberPK.java
@@ -1,5 +1,7 @@
-package com.sixman.roomus.product.command.domain.model;
+package com.sixman.roomus.rooms.command.domain.model.vo;
 
+
+import com.sixman.roomus.rooms.command.domain.model.Room;
 import lombok.*;
 
 import javax.persistence.Column;
@@ -15,18 +17,18 @@ import java.util.Objects;
 @AllArgsConstructor
 @NoArgsConstructor
 @ToString
-public class ProductLikesMemberPK implements Serializable {
+public class RoomLikesMemberPK implements Serializable {
 
     @ManyToOne
-    @JoinColumn(name = "PRODUCT_NO")
-    private Product product;
+    @JoinColumn(name = "ROOM_NO")
+    private Room room;
 
     @Column(name = "MEMBER_NO")
     private int memberNo;
 
     @Override
     public int hashCode() {
-        return Objects.hash(product.hashCode(), memberNo);
+        return Objects.hash(room.hashCode(), memberNo);
     }
 
     @Override
@@ -37,7 +39,7 @@ public class ProductLikesMemberPK implements Serializable {
         if (obj == null || getClass() != obj.getClass()) {
             return false;
         }
-        ProductLikesMemberPK productLikesMemberPK = (ProductLikesMemberPK) obj;
-        return product == productLikesMemberPK.getProduct() && Objects.equals(memberNo, productLikesMemberPK.getMemberNo());
+        RoomLikesMemberPK roomLikesMemberPK = (RoomLikesMemberPK) obj;
+        return this.room == roomLikesMemberPK.getRoom() && this.memberNo == roomLikesMemberPK.getMemberNo();
     }
 }
