+++ src/main/java/com/sixman/roomus/rooms/query/model/FurnitureArrangementData.java
@@ -1,7 +1,7 @@
 package com.sixman.roomus.rooms.query.model;
 
 
-import com.sixman.roomus.rooms.command.domain.model.Vector3;
+import com.sixman.roomus.rooms.command.domain.model.vo.Vector3;
 import lombok.*;
 
 import javax.persistence.*;
